==> notebooks/1-data-transformation.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import pickle
import re

import nltk
nltk.download('stopwords');

# 1. Read in Raw Data

train = pd.read_csv('../data/raw/train.csv')
test = pd.read_csv('../data/raw/valid.csv')

train.shape[0], test.shape[0]

train.head(2)

# train.loc[0].CreationDate

# all_data = pd.concat([train, test]).reset_index(drop = True)

# all_data.shape

# train.query('Y == "HQ"').shape[0], test.query('Y == "HQ"').shape[0]

# train.query('Y == "LQ_CLOSE"').shape[0], test.query('Y == "LQ_CLOSE"').shape[0]

# train.query('Y == "LQ_EDIT"').shape[0], test.query('Y == "LQ_CLOSE"').shape[0]

# 2. Data Cleaning

## 2.1 Check NaN

train.isna().sum()

test.isna().sum()

## 2.2 Clean Label

label_dict = {
    'LQ_CLOSE': 0,
    'LQ_EDIT': 1,
    'HQ': 2
}

train['cleaned_y'] = train.Y.apply(lambda x: label_dict[x])
test['cleaned_y'] = test.Y.apply(lambda x: label_dict[x])

## 2.3 Clean Title

# title
train.Title = train.Title.apply(lambda x: x.lower())
test.Title = test.Title.apply(lambda x: x.lower())

from nltk import word_tokenize
from nltk.corpus import stopwords
stop_words = stopwords.words('english')

def remove_stopwords(string):
    word_list = [word.lower() for word in string.split()]
    stopwords_list = list(stopwords.words("english"))
    for word in word_list:
        if word in stopwords_list:
            word_list.remove(word)
    return ' '.join(word_list)

def clean_text(text):
    '''
    Acknowledgement: this function is inspired by the following post:
        https://www.kaggle.com/anmolkumar/stack-overflow-eda-bert-model-accuracy-87-6
    '''
    text = re.sub('\\n', ' ', text)
    text = re.sub('\W', ' ', text)
    text = re.sub(r'https\s+|www.\s+', r'', text)
    text = re.sub(r'http\s+|www.\s+',r'', text)
    text = re.sub(r'\s+[a-zA-Z]\s+',' ', text)
    text = re.sub(r'\^[a-zA-Z]\s+',' ', text)
    text = text.lower()
    
    text = re.sub(r"\’", "\'", text)
    text = re.sub(r"won\'t", "will not", text)
    text = re.sub(r"can\'t", "can not", text)
    text = re.sub(r"don\'t", "do not", text)
    text = re.sub(r"dont", "do not", text)
    text = re.sub(r"n\’t", " not", text)
    text = re.sub(r"n\'t", " not", text)
    text = re.sub(r"\'re", " are", text)
    text = re.sub(r"\'s", " is", text)
    text = re.sub(r"\’d", " would", text)
    text = re.sub(r"\d", " would", text)
    text = re.sub(r"\'ll", " will", text)
    text = re.sub(r"\'t", " not", text)
    text = re.sub(r"\'ve", " have", text)
    text = re.sub(r"\'m", " am", text)
    text = re.sub(r"\n", "", text)
    text = re.sub(r"\r", "", text)
    text = re.sub(r"[0-9]", "digit", text)
    text = re.sub(r"\'", "", text)
    text = re.sub(r"\"", "", text)
    text = re.sub(r'[?|!|\'|"|#]',r'', text)
    text = re.sub(r'[.|,|)|(|\|/]',r' ', text)
    text = remove_stopwords(text)
    
    return text

train.Title = train.Title.apply(lambda x: clean_text(x))
test.Title = test.Title.apply(lambda x: clean_text(x))

## 2.4 Clean Body

import re

train['body_notag'] = train.Body.apply(lambda s: re.sub('<[^>]+>', '', s))
test['body_notag'] = test.Body.apply(lambda s: re.sub('<[^>]+>', '', s))

train.body_notag = train.body_notag.apply(lambda x: clean_text(x))
test.body_notag = test.body_notag.apply(lambda x: clean_text(x))

## 2.5 Combine Body and Title

train['all_text'] = train.Title + ' ' + train.body_notag
test['all_text'] = test.Title + ' ' + test.body_notag

# 3. Feature Engineering

## 3.1. Length of the Title & Question

train['title_length'] = train.Title.apply(lambda x: len(x.split()))
test['title_length'] = test.Title.apply(lambda x: len(x.split()))

train['body_length'] = train.body_notag.apply(lambda x: len(x.split()))
test['body_length'] = test.body_notag.apply(lambda x: len(x.split()))

# 4. Train-Val-Test Split

from sklearn.model_selection import train_test_split

# 45000 * 0.22222222

train_df, val_df, _, _ = train_test_split(train, train['Y'], test_size=0.22222222, random_state=42)

train_df.shape, val_df.shape

pickle.dump(train_df, open('../data/processed/train_df.pkl', 'wb'))
pickle.dump(val_df, open('../data/processed/val_df.pkl', 'wb'))
pickle.dump(test, open('../data/processed/test_df.pkl', 'wb'))
