==> notebooks/3-benchmark.ipynb <==
# 0. Overview
- This notebook handles the exploratory data analysis of the given dataset.


import pickle
import pandas as pd
import numpy as np
from tqdm import tqdm

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

train_df = pickle.load(open('../data/processed/train_df.pkl', 'rb'))
val_df = pickle.load(open('../data/processed/val_df.pkl', 'rb'))
test_df = pickle.load(open('../data/processed/test_df.pkl', 'rb'))

X_train = train_df[['title_length', 'body_length']]
X_val = val_df[['title_length', 'body_length']]
X_test = test_df[['title_length', 'body_length']]

y_train = train_df.cleaned_y.values
y_val = val_df.cleaned_y.values
y_test = test_df.cleaned_y.values

all_rows = []
for c in tqdm([0.001, 0.01, 0.1, 1, 10, 100, 1000]):
    clf = LogisticRegression(random_state = 0).fit(X_train, y_train)
    train_acc = accuracy_score(y_train, clf.predict(X_train))
    val_acc = accuracy_score(y_val, clf.predict(X_val))
    all_rows.append({
        'c': c,
        'train_acc': round(train_acc, 2),
        'val_acc': round(val_acc, 2)
    })

pd.DataFrame(all_rows)

accuracy_score(y_test, clf.predict(X_test))

test_prediction = clf.predict(X_test)

pickle.dump(test_prediction, open('../data/baseline_pred.pkl', 'wb'))
