==> notebooks/4-huggingface.ipynb <==
!pip install "sagemaker>=2.31.0" "transformers==4.6.1" "datasets[s3]==1.5.0" --upgrade

import pickle
import pandas as pd
import numpy as np
from tqdm import tqdm
import os
import datasets

data_dir = '../data/bert' # The folder we will use for storing data
os.makedirs(data_dir, exist_ok=True)  # ensure cache directory exists

cache_dir = os.path.join("../cache", "bert")  # where to store cache files
os.makedirs(cache_dir, exist_ok=True)  # ensure cache directory exists

datasets.__version__

# 1. Preparing Data

## 1.1. Loading Dataset

from datasets import load_dataset

# train_df = pickle.load(open('../data/processed/train_df.pkl', 'rb'))
# val_df = pickle.load(open('../data/processed/val_df.pkl', 'rb'))
# test_df = pickle.load(open('../data/processed/test_df.pkl', 'rb'))

# train_df = pd.concat([train_df, val_df]).reset_index(drop = True)
# val_df = None

# train_df = train_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})
# # val_df = val_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})
# test_df = test_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})

# train_df[['label', 'sentence']].to_csv(os.path.join(data_dir, 'train_bert.csv'), index = False, header = True)
# # val_df[['label', 'sentence']].to_csv(os.path.join(data_dir, 'val_bert.csv'), index = False, header = True)
# test_df[['label', 'sentence']].to_csv(os.path.join(data_dir, 'test_bert.csv'), index = False, header = True)

train_dataset = load_dataset('csv', data_files=os.path.join(data_dir, 'train_bert.csv'))['train']
test_dataset = load_dataset('csv', data_files=os.path.join(data_dir, 'test_bert.csv'))['train']

test_dataset

## 1.2 Process Dataset

# from datasets import load_dataset
from transformers import AutoTokenizer

# tokenizer used in preprocessing
tokenizer_name = 'distilbert-base-uncased'

# dataset used
# dataset_name = 'imdb'

# s3 key prefix for the data
s3_prefix = 'sagemaker/stackoverflow-question-quality'

# download tokenizer
tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)

# tokenizer helper function
def tokenize(batch):
    return tokenizer(batch['sentence'], padding='max_length', truncation=True)

train_dataset

# tokenize dataset
train_dataset = train_dataset.map(tokenize, batched=True)
test_dataset = test_dataset.map(tokenize, batched=True)

# set format for pytorch
train_dataset =  train_dataset.rename_column("label", "labels")
train_dataset.set_format('torch', columns=['input_ids', 'attention_mask', 'labels'])
test_dataset = test_dataset.rename_column("label", "labels")
test_dataset.set_format('torch', columns=['input_ids', 'attention_mask', 'labels'])

train_dataset

## 1.3 Uploading Data

import sagemaker

sess = sagemaker.Session()
# sagemaker session bucket -> used for uploading data, models and logs
# sagemaker will automatically create this bucket if it not exists
sagemaker_session_bucket=None
if sagemaker_session_bucket is None and sess is not None:
    # set to default bucket if a bucket name is not given
    sagemaker_session_bucket = sess.default_bucket()

role = sagemaker.get_execution_role()
sess = sagemaker.Session(default_bucket=sagemaker_session_bucket)

print(f"sagemaker role arn: {role}")
print(f"sagemaker bucket: {sess.default_bucket()}")
print(f"sagemaker session region: {sess.boto_region_name}")

import botocore
from datasets.filesystems import S3FileSystem

s3 = S3FileSystem()  

# save train_dataset to s3
training_input_path = f's3://{sess.default_bucket()}/{s3_prefix}/train_bert'
train_dataset.save_to_disk(training_input_path,fs=s3)

# save test_dataset to s3
test_input_path = f's3://{sess.default_bucket()}/{s3_prefix}/test_bert'
test_dataset.save_to_disk(test_input_path,fs=s3)

test_input_path

# from datasets import load_from_disk

# t = load_from_disk(test_input_path)

# # load dataset
# dataset = load_dataset(dataset_name)

# 2. Create Model

## 2.1 Model Training

training_input_path

from sagemaker.huggingface import HuggingFace

# hyperparameters, which are passed into the training job
hyperparameters={'epochs': 1,
                 'per_device_train_batch_size': 2,
                 'model_name': 'bert-base-uncased'
                 }

# create the Estimator
huggingface_estimator = HuggingFace(
        entry_point='huggingface_train.py',
        source_dir='../src_bert',
        instance_type='ml.p2.xlarge',
        instance_count=1,
        role=role,
        transformers_version='4.4',
        pytorch_version='1.6',
        py_version='py36',
        hyperparameters = hyperparameters
)

# starting the train job with our uploaded datasets as input
huggingface_estimator.fit({'train': training_input_path, 'test': test_input_path})

## 2.2 Model Evaluation

# container image used for training job
print(f"container image used for training job: \n{huggingface_estimator.image_uri}\n")

# s3 uri where the trained model is located
print(f"s3 uri where the trained model is located: \n{huggingface_estimator.model_data}\n")

# latest training job name for this estimator
print(f"latest training job name for this estimator: \n{huggingface_estimator.latest_training_job.name}\n")

huggingface_transformer = huggingface_estimator.transformer(instance_count = 1, instance_type = 'ml.m4.xlarge')

# xgb_transformer.transform(test_location, content_type='text/csv', split_type='Line')

# evaluate result:
# open(os.path.join(args.output_data_dir, "eval_results.txt"), "w")
