==> notebooks/6-sklearn.ipynb <==
import pickle
import pandas as pd
import numpy as np
from tqdm import tqdm
import os

from sklearn.metrics import accuracy_score

data_dir = '../data/sklearn' # The folder we will use for storing data
os.makedirs(data_dir, exist_ok=True)  # ensure cache directory exists

cache_dir = os.path.join("../cache", "lstm")  # where to store cache files
os.makedirs(cache_dir, exist_ok=True)  # ensure cache directory exists

# 1. Data Prep

## 1.1 Load Data

train_df = pickle.load(open('../data/processed/train_df.pkl', 'rb'))
val_df = pickle.load(open('../data/processed/val_df.pkl', 'rb'))
test_df = pickle.load(open('../data/processed/test_df.pkl', 'rb'))

train_df = train_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})
val_df = val_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})
test_df = test_df.rename(columns = {'cleaned_y': 'label', 'all_text': 'sentence'})

train_df = pd.concat([train_df, val_df]).reset_index(drop = True)
val_df = None

test_y = test_df.label.values

## 1.2 Question to List of Words

import nltk
from nltk.corpus import stopwords
from nltk.stem.porter import *

import os
import re
from bs4 import BeautifulSoup

def question_to_words(question):
    nltk.download("stopwords", quiet=True)
    stemmer = PorterStemmer()
    #print(review)
    #text = BeautifulSoup(review, "html.parser").get_text() # Remove HTML tags
    text = re.sub(r"[^a-zA-Z0-9]", " ", question.lower()) # Convert to lower case
    words = text.split() # Split string into words
    words = [w for w in words if w not in stopwords.words("english")] # Remove stopwords
    words = [PorterStemmer().stem(w) for w in words] # stem
    
    return words

import pickle

def preprocess_data(data_train, data_test, labels_train, labels_test,
                    cache_dir=cache_dir, cache_file="preprocessed_data.pkl"):
    """Convert each review to words; read from cache if available."""

    # If cache_file is not None, try to read from it first
    cache_data = None
    if cache_file is not None:
        try:
            with open(os.path.join(cache_dir, cache_file), "rb") as f:
                cache_data = pickle.load(f)
            print("Read preprocessed data from cache file:", cache_file)
        except:
            pass  # unable to read from cache, but that's okay
    
    # If cache is missing, then do the heavy lifting
    if cache_data is None:
        # Preprocess training and test data to obtain words for each review
        words_train = [question_to_words(question) for question in tqdm(data_train)]
        words_test = [question_to_words(question) for question in tqdm(data_test)]
        
        # Write to cache file for future runs
        if cache_file is not None:
            cache_data = dict(words_train=words_train, words_test=words_test,
                              labels_train=labels_train, labels_test=labels_test)
            with open(os.path.join(cache_dir, cache_file), "wb") as f:
                pickle.dump(cache_data, f)
            print("Wrote preprocessed data to cache file:", cache_file)
    else:
        # Unpack data loaded from cache file
        words_train, words_test, labels_train, labels_test = (cache_data['words_train'],
                cache_data['words_test'], cache_data['labels_train'], cache_data['labels_test'])
    
    return words_train, words_test, labels_train, labels_test

train_X, train_y = train_df.sentence.values, train_df.label.values
test_X, test_y = test_df.sentence.values, test_df.label.values

train_X, test_X, train_y, test_y = preprocess_data(train_X, test_X, train_y, test_y)

train_df = test_df = None

# train_X.shape

## 1.3 Extract BOW features

import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
# from sklearn.externals 
import joblib
# joblib is an enhanced version of pickle that is more efficient for storing NumPy arrays

# with open(os.path.join(cache_dir, "bow_features.pkl"), "rb") as f:
#     cache_data = joblib.load(f)

def extract_BoW_features(words_train, words_test, vocabulary_size=5000,
                         cache_dir=cache_dir, cache_file="bow_features.pkl"):
    """Extract Bag-of-Words for a given set of documents, already preprocessed into words."""
    
    # If cache_file is not None, try to read from it first
    cache_data = None
    if cache_file is not None:
        try:
            with open(os.path.join(cache_dir, cache_file), "rb") as f:
                cache_data = joblib.load(f)
            print("Read features from cache file:", cache_file)
        except:
            pass  # unable to read from cache, but that's okay
    
    # If cache is missing, then do the heavy lifting
    if cache_data is None:
        # Fit a vectorizer to training documents and use it to transform them
        # NOTE: Training documents have already been preprocessed and tokenized into words;
        #       pass in dummy functions to skip those steps, e.g. preprocessor=lambda x: x
        vectorizer = CountVectorizer(max_features=vocabulary_size,
                preprocessor=lambda x: x, tokenizer=lambda x: x)  # already preprocessed
        features_train = vectorizer.fit_transform(words_train).toarray()

        # Apply the same vectorizer to transform the test documents (ignore unknown words)
        features_test = vectorizer.transform(words_test).toarray()
        
        # NOTE: Remember to convert the features using .toarray() for a compact representation
        
        # Write to cache file for future runs (store vocabulary as well)
        if cache_file is not None:
            vocabulary = vectorizer.vocabulary_
            cache_data = dict(features_train=features_train, features_test=features_test,
                             vocabulary=vocabulary)
            with open(os.path.join(cache_dir, cache_file), "wb") as f:
                joblib.dump(cache_data, f)
            print("Wrote features to cache file:", cache_file)
    else:
        # Unpack data loaded from cache file
        features_train, features_test, vocabulary = (cache_data['features_train'],
                cache_data['features_test'], cache_data['vocabulary'])
    
    # Return both the extracted features as well as the vocabulary
    return features_train, features_test, vocabulary

# Extract Bag of Words features for both training and test datasets
train_X, test_X, vocabulary = extract_BoW_features(train_X, test_X)

- **Test set**

pd.DataFrame(test_X).to_csv(os.path.join(data_dir, 'test_sklearn.csv'), header=False, index=False)
test_X = vocabulary = None

- **Validation Set**

val_idx = np.random.choice(range(len(train_y)), size=10000, replace=False)

val_X = pd.DataFrame(train_X[val_idx])
val_y = pd.DataFrame(train_y[val_idx])

val_X.shape, val_y.shape

pd.concat([val_y, val_X], axis=1).to_csv(os.path.join(data_dir, 'val_sklearn.csv'), header=False, index=False)
val_y = val_X = None

- **Train Set**

train_idx = [i for i in range(len(train_y)) if i not in val_idx]

train_X = pd.DataFrame(train_X[train_idx])
train_y = pd.DataFrame(train_y[train_idx])

train_X.shape, train_y.shape

pd.concat([train_y, train_X], axis=1).to_csv(os.path.join(data_dir, 'train_sklearn.csv'), header=False, index=False)

test_X = train_X = val_X = train_y = val_y = None

## 1.4 Upload to S3

import sagemaker

session = sagemaker.Session() # Store the current SageMaker session
bucket = sagemaker.Session().default_bucket()

prefix = 'sagemaker/stackoverflow-question-quality'

test_location = session.upload_data(os.path.join(data_dir, 'test_sklearn.csv'), key_prefix=prefix)
val_location = session.upload_data(os.path.join(data_dir, 'val_sklearn.csv'), key_prefix=prefix)
train_location = session.upload_data(os.path.join(data_dir, 'train_sklearn.csv'), key_prefix=prefix)

# test_location = session.upload_data(os.path.join(data_dir, 'test_sklearn.csv'), key_prefix=prefix)

# t = pd.read_csv(data_dir + '/test_sklearn.csv', header = None)
# t.head()

# t[5].unique()

# t.shape

# t = pd.read_csv(data_dir + '/val_sklearn.csv', header = None)
# t.head()

# 2. Model Training

sagemaker.__version__

# pip install -U sagemaker

import sagemaker
from sagemaker import image_uris
from sagemaker.inputs import TrainingInput
from sagemaker import get_execution_role

role = get_execution_role()
container = image_uris.retrieve(framework='xgboost',region='us-east-1',version='1.2-1')

xgb = sagemaker.estimator.Estimator(container, # The location of the container we wish to use
                                    role,                                    # What is our current IAM Role
                                    train_instance_count=1,                  # How many compute instances
                                    train_instance_type='ml.m4.xlarge',      # What kind of compute instances
                                    output_path='s3://{}/{}/sklearn_output'.format(session.default_bucket(), prefix),
                                    sagemaker_session=session)

xgb.set_hyperparameters(max_depth=11,
                        eta=0.2622829413333343,
                        gamma=1.0375026465214707,
                        min_child_weight=3,
                        subsample=0.8544514211613925,
                        #silent=0,
                        objective='multi:softmax',
                        num_class=3,
                        early_stopping_rounds=10,
                        num_round=500)

train_input = TrainingInput("s3://{}/{}/{}".format(bucket, prefix, 'train_sklearn.csv'), content_type='csv')
validation_input = TrainingInput("s3://{}/{}/{}".format(bucket, prefix, 'val_sklearn.csv'), content_type='csv')

"s3://{}/{}/{}".format(bucket, prefix, 'train_sklearn.csv')# validation_input

xgb.fit({'train': train_input, 'validation': validation_input})

# 3. Model Evaluation

# "s3://{}/{}/{}".format(bucket, prefix, 'val_sklearn.csv')
test_location = 's3://sagemaker-us-east-1-997893341280/sagemaker/stackoverflow-question-quality/test_sklearn.csv'

xgb_transformer = xgb.transformer(instance_count = 1, instance_type = 'ml.m4.xlarge')

xgb_transformer.transform(test_location, content_type='text/csv', split_type='Line')

# test_location

# xgb_transformer.wait()
xgb_transformer.output_path

# !aws s3 cp --recursive $xgb_transformer.output_path $data_dir

predictions = pd.read_csv(os.path.join(xgb_transformer.output_path, 'test_sklearn.csv.out'), header=None)
predictions = [round(num) for num in predictions.squeeze().values]

accuracy_score(test_y, predictions)

test_y

# !$xgb_transformer.output_path
# bucket

# !aws s3 cp
# data_dir

# 4. Hyperparameter Tuning

from sagemaker.tuner import IntegerParameter, ContinuousParameter, HyperparameterTuner

xgb = sagemaker.estimator.Estimator(container, # The location of the container we wish to use
                                    role,                                    # What is our current IAM Role
                                    train_instance_count=1,                  # How many compute instances
                                    train_instance_type='ml.m4.xlarge',      # What kind of compute instances
                                    output_path='s3://{}/{}/sklearn_output'.format(session.default_bucket(), prefix),
                                    sagemaker_session=session)

xgb.set_hyperparameters(max_depth=5,
                        eta=0.2,
                        gamma=4,
                        min_child_weight=6,
                        subsample=0.8,
                        #silent=0,
                        objective='multi:softmax',
                        num_class=3,
                        early_stopping_rounds=10,
                        num_round=500)

xgb_hyperparameter_tuner = HyperparameterTuner(estimator = xgb, # The estimator object to use as the basis for the training jobs.
                                               objective_metric_name = 'validation:merror', # The metric used to compare trained models.
                                               objective_type = 'Minimize', # Whether we wish to minimize or maximize the metric.
                                               max_jobs = 6, # The total number of models to train
                                               max_parallel_jobs = 3, # The number of models to train in parallel
                                               hyperparameter_ranges = {
                                                    'max_depth': IntegerParameter(3, 12),
                                                    'eta'      : ContinuousParameter(0.05, 0.5),
                                                    'min_child_weight': IntegerParameter(2, 8),
                                                    'subsample': ContinuousParameter(0.5, 0.9),
                                                    'gamma': ContinuousParameter(0, 10),
                                               })

train_input = TrainingInput("s3://{}/{}/{}".format(bucket, prefix, 'train_sklearn.csv'), content_type='csv')
validation_input = TrainingInput("s3://{}/{}/{}".format(bucket, prefix, 'val_sklearn.csv'), content_type='csv')

xgb_hyperparameter_tuner.fit({'train': train_input, 'validation': validation_input})

test_location = 's3://sagemaker-us-east-1-997893341280/sagemaker/stackoverflow-question-quality/test_sklearn.csv'
# best_training_job_location = 's3://sagemaker-us-east-1-997893341280/sagemaker/stackoverflow-question-quality/sklearn_output/sagemaker-xgboost-210708-0851-006-1c702236/output'

xgb_attached = sagemaker.estimator.Estimator.attach('sagemaker-xgboost-2021-07-08-14-02-27-286')

xgb_transformer = xgb_attached.transformer(instance_count = 1, instance_type = 'ml.m4.xlarge')

# test_input = TrainingInput("s3://{}/{}/{}".format(bucket, prefix, 'test_sklearn.csv'), content_type='csv')

xgb_transformer.transform(test_location, content_type='text/csv', split_type='Line')

# xgb_transformer.wait()

# pickle.dump(xgb_hyperparameter_tuner, open('../data/sklearn/xgb_hyperparameter_tuner.pkl', 'wb'))

# !aws s3 cp --recursive $xgb_transformer.output_path $data_dir

predictions = pd.read_csv(os.path.join(data_dir, 'test_sklearn.csv.out'), header=None)
predictions = [round(num) for num in predictions.squeeze().values]

accuracy_score(test_y, predictions)
