==> notebooks/7-result-analysis.ipynb <==
import pandas as pd
from sklearn.metrics import accuracy_score
import pickle

# 0. Read in test output

out_1 = pickle.load(open('../data/baseline_pred.pkl', 'rb'))
# out_2 = pd.read_csv('../data/sklearn/test_sklearn_out_2.csv', header = None)
test_df = pickle.load(open('../data/processed/test_df.pkl', 'rb'))

# test_df

test_y = test_df.cleaned_y.values
test_df.shape

accuracy_score(test_y, out_1) # baseline

accuracy_score(test_y, out_2.values.flatten()) # xgboost

# 1. Baseline
## 1.1 Performance Distribution

pred_y = out_1

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 0:
        total_points += 1
        if pred_y[i] == 0:
            total_correct += 1
acc_class_0 = total_correct / total_points
acc_class_0

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 1:
        total_points += 1
        if pred_y[i] == 1:
            total_correct += 1
acc_class_1 = total_correct / total_points
acc_class_1

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 2:
        total_points += 1
        if pred_y[i] == 2:
            total_correct += 1
acc_class_2 = total_correct / total_points
acc_class_2

## 1.2 One-VS-All

test_y_ova = []
for y in test_y:
    if y != 2:
        test_y_ova.append(0)
    else:
        test_y_ova.append(1)

pred_y_ova = []
for y in pred_y:
    if y != 2:
        pred_y_ova.append(0)
    else:
        pred_y_ova.append(1)

accuracy_score(test_y_ova, pred_y_ova)

test_y_ova = []
for y in test_y:
    if y != 0:
        test_y_ova.append(1)
    else:
        test_y_ova.append(0)
        
pred_y_ova = []
for y in pred_y:
    if y != 0:
        pred_y_ova.append(1)
    else:
        pred_y_ova.append(0)
accuracy_score(test_y_ova, pred_y_ova)

# 2. XGBoost Model 
## 2.1 Performance Distribution

pred_y = out_2.values.flatten()

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 0:
        total_points += 1
        if pred_y[i] == 0:
            total_correct += 1
acc_class_0 = total_correct / total_points
acc_class_0

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 1:
        total_points += 1
        if pred_y[i] == 1:
            total_correct += 1
acc_class_1 = total_correct / total_points
acc_class_1

total_points = 0
total_correct = 0
for i in range(len(pred_y)):
    if test_y[i] == 2:
        total_points += 1
        if pred_y[i] == 2:
            total_correct += 1
acc_class_2 = total_correct / total_points
acc_class_2

# 2.2. One-VS-All

test_y_ova = []
for y in test_y:
    if y != 2:
        test_y_ova.append(0)
    else:
        test_y_ova.append(1)

pred_y_ova = []
for y in pred_y:
    if y != 2:
        pred_y_ova.append(0)
    else:
        pred_y_ova.append(1)

accuracy_score(test_y_ova, pred_y_ova)

test_y_ova = []
for y in test_y:
    if y != 0:
        test_y_ova.append(1)
    else:
        test_y_ova.append(0)
        
pred_y_ova = []
for y in pred_y:
    if y != 0:
        pred_y_ova.append(1)
    else:
        pred_y_ova.append(0)
accuracy_score(test_y_ova, pred_y_ova)
